removed code implementation for feature toggling
